requirements.txt
python-telegram-bot[job-queue]>=20.0
python-dotenv
feedparser
requests
google-genai>=1.0.0
fastapi
uvicorn[standard]
pydantic>=2.0
beautifulsoup4
lxml
pandas
